MapField Done

+++ DAL/Data Access Objects/IdeaFieldsDao.cs
@@ -6,8 +6,8 @@
         public int IdeaId { get; set; }
         public string FieldText { get; set; }
         public string FieldStrings { get; set; }
-        public float LocationX { get; set; }
-        public float LocationY { get; set; }
+        public double LocationX { get; set; }
+        public double LocationY { get; set; }
         public string Url { get; set; }
         public byte[] UploadedImage { get; set; }
         public byte[] UploadedMedia { get; set; }
